Moved function from StatusCache to Status Added Taurenkey's logic under Tanks for better handling
