+++ WrathCombo/CustomCombo/Functions/Status.cs
@@ -1,5 +1,6 @@
 ﻿using Dalamud.Game.ClientState.Objects.SubKinds;
 using Dalamud.Game.ClientState.Objects.Types;
+using ECommons.DalamudServices;
 using ECommons.GameFunctions;
 using FFXIVClientStructs.FFXIV.Client.Game;
 using System.Linq;
@@ -129,14 +130,133 @@ namespace WrathCombo.CustomComboNS.Functions
         /// <returns></returns>
         public static bool HasBeneficialStatus(IGameObject? target) => StatusCache.HasBeneficialStatus(target);
 
-        public static bool HasPhantomDispelStatus(IGameObject? target) => StatusCache.HasDamageUp(target) || StatusCache.HasEvasionUp(target) || HasStatusEffect(4355, target) || StatusCache.TargetIsInvincible(target);
+        public static bool HasPhantomDispelStatus(IGameObject? target) => StatusCache.HasDamageUp(target) || StatusCache.HasEvasionUp(target) || HasStatusEffect(4355, target) || TargetIsInvincible(target);
 
         /// <summary>
         /// Checks if the target is invincible due to status effects or encounter-specific mechanics.
         /// </summary>
         /// <param name="target">The game object to check.</param>
         /// <returns>True if the target is invincible; otherwise, false.</returns>
-        public static bool TargetIsInvincible(IGameObject? target) => StatusCache.TargetIsInvincible(target);
+        public static bool TargetIsInvincible(IGameObject? target)
+        {
+            if (target is not IBattleChara tar)
+                return false;
+
+            // Turn Target's status to uint hashset
+            var targetStatuses = tar.StatusList.Select(s => s.StatusId).ToHashSet();
+            uint targetID = tar.DataId;
+
+            switch (Svc.ClientState.TerritoryType)
+            {
+                case 174:   // Labyrinth of the Ancients
+                    // Thanatos, Spooky Ghosts Only
+                    if (targetID is 2350) return !HasStatusEffect(398);
+
+                    // Allagan Bomb
+                    if (targetID is 2407) return NumberOfEnemiesInRange(30) > 1;
+
+                    return false;
+                case 1248:  // Jeuno 1 Ark Angels
+                    // ArkAngel HM = 1804
+                    if (targetID is 18049 && HasStatusEffect(4410, tar, true)) return true;
+
+                    // ArkAngel MR = 18051 (A)
+                    // ArkAngel GK = 18053 (B)
+                    // ArkAngel TT = 18052 (C)
+                    if (targetID is 18051 or 18052 or 18053)
+                    {
+                        if (HasStatusEffect(4192)) return targetID != 18051; // Alliance A Red Epic
+                        if (HasStatusEffect(4194)) return targetID != 18053; // Alliance B Yellow Fated
+                        if (HasStatusEffect(4196)) return targetID != 18052; // Alliance C Blue Vaunted
+                    }
+                    return false;
+                case 917:   //Puppet's Bunker, Flight Mechs
+                    // 724P Alpha = 11792 (A)
+                    // 767P Beta  = 11793 (B)
+                    // 772P Chi   = 11794 (C)
+                    if (targetID is 11792 or 11793 or 11794)
+                    {
+                        if (HasStatusEffect(2288)) return targetID != 11792;
+                        if (HasStatusEffect(2289)) return targetID != 11793;
+                        if (HasStatusEffect(2290)) return targetID != 11794;
+                    }
+                    return false;
+                case 966: //The Tower at Paradigm's Breach, Hansel & Gretel
+                          // Hansel = 12709
+                          // Gretel = 12708
+                          // 608 Directional Parry
+                          // 2538 Strong of Shield
+                          // 2539 Stronger Together
+
+                    if (targetID is 12709 or 12708)
+                    {
+                        bool Tank = (LocalPlayer!).GetRole() is CombatRole.Tank;
+
+                        // Non Tanks should just ignore parrying boss(s)
+                        if (!Tank)
+                        {
+                            if (HasStatusEffect(680, tar)) return true;
+                        }
+                        else // Better Tank logic to ignore the other boss, from Taurenkey
+                        {
+                            int[] bosses = [12709, 12708];
+                            var otherBoss = Svc.Objects.FirstOrDefault(x => x.DataId == bosses.FirstOrDefault(x => x != targetID) && !x.IsDead);
+                            bool inFrontOfBoss = AngleToTarget(tar) == AttackAngle.Front;
+                            bool bossHasStatus = HasStatusEffect(680, tar);
+
+                            if (inFrontOfBoss && bossHasStatus) //Scenario 1: Boss has buff, you're in front, it's good to attack
+                                return false;
+
+                            if (!bossHasStatus) //Scenario 2: The boss doesn't have a buff, we only want to target it if we're not in front of the boss with the buff
+                            {
+                                if (otherBoss is not null)
+                                {
+                                    if (HasStatusEffect(680, otherBoss) && AngleToTarget(otherBoss) == AttackAngle.Front)
+                                        return true;
+
+                                    return false;
+                                }
+                            }
+
+                            if (bossHasStatus) //Scenario 3: Boss has buff, you're not in front, it's not good to attack
+                                return true;
+                        }
+                    }
+                    return false;
+
+                case 801 or 805 or 1122: //Interdimensional Rift (Omega 12 / Alphascape 4), Regular/Savage?/Ultimate?
+                    // Omega-M = 9339
+                    // Omega-F = 9340
+                    if (targetID is 9339 or 9340) //numbers are for Regular
+                    {
+                        if (HasStatusEffect(1660)) return targetID == 9339; // Packet Filter M
+                        if (HasStatusEffect(1661)) return targetID == 9340; // Packet Filter F
+                        if (targetID is 9340) return HasStatusEffect(671, tar, true); // F being covered by M
+                    }
+
+                    //Savage/Ultimate? Not sure which omega fight uses 3499 and 3500.
+                    //Also, SE, why use a new Omega-M status and reuse the old Omega-F? -_-'
+                    //Wonder if targetIDs are the same......
+                    if ((tar.StatusList.Any(x => x.StatusId == 3454) && HasStatusEffect(3499)) ||
+                        (tar.StatusList.Any(x => x.StatusId == 1675) && HasStatusEffect(3500)))
+                        return true;
+
+                    //Check for any ol invincibility
+                    if (StatusCache.CompareLists(StatusCache.InvincibleStatuses, targetStatuses)) return true;
+
+                    return false;
+                case 952:  //ToZ final boss (technically not invincible)
+                    if (targetID is (13298 or 13299) && Svc.Objects.Any(y => y.DataId is 13297 && !y.IsDead))
+                        return true;
+
+                    return false;
+            }
+
+            // General invincibility check
+            // Due to large size of InvincibleStatuses, best to check process this way
+            if (StatusCache.CompareLists(StatusCache.InvincibleStatuses, targetStatuses)) return true;
+            return false;
+        }
 
         /// <summary>
         /// Checks if a target has the max number of entries in their status list.
